services/worker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-firestore==2.14.0
google-cloud-pubsub==2.27.2
google-cloud-logging==3.11.0
